Revert "add unit tests for checkRepositoryPermissions (#1016)"
This reverts commit cf26e5da0b39206dac9889b59bd3abbef3fee181.
